[BE/REFACTOR] BIGDECIMAL > double로 수정

+++ backend/src/main/java/com/prana/backend/calendar/controller/response/WeeklyReportResponseDTO.java
@@ -54,8 +54,8 @@ public class WeeklyReportResponseDTO {
         private int month;
         private int week;
         private int time;
-        private BigDecimal accuracy;
-        private BigDecimal bmi;
+        private double accuracy;
+        private double bmi;
 
     }
 }
